Fix unit test and move FileDownloaderTest
